Reorganized and renamed files
