--- client/src/components/Account/index.js
@@ -1,7 +1,0 @@
-import React from 'react';
-const Account = () => (
-  <div>
-    <h1>Account</h1>
-  </div>
-);
-export default Account;
