--- client/src/components/Admin/index.js
@@ -1,7 +1,0 @@
-import React from 'react';
-const Admin = () => (
-  <div>
-    <h1>Admin</h1>
-  </div>
-);
-export default Admin;
