--- client/src/components/Home/index.js
@@ -1,7 +1,0 @@
-import React from 'react';
-const Home = () => (
-  <div>
-    <h1>Home</h1>
-  </div>
-);
-export default Home;
