--- client/src/components/PasswordForget/index.js
@@ -1,7 +1,0 @@
-import React from 'react';
-const PasswordForget = () => (
-  <div>
-    <h1>PasswordForget</h1>
-  </div>
-);
-export default PasswordForget;
